+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Add process_submissions().
 
+* Remove the ability to save answers as either RDS or PDF files.
+
 # tutorial.helpers 0.2.7
 
 * Get test cases for `format_tutorial()` working again.